Factor out a list of targets that don't support GDB
There's a lot of targets for which we can build binutils/ld/gas but
not build GDB.  Instead of having a separate CONFIGURE_* variable
defined for each of these targets, I now create a list of targets, and
use this inside a $(foreach ....) to automate the creation of the
CONFIGURE_* variables.

This reduces the amount of duplication, and makes things easier to
understand.