Split some of the configure/build rules into definitions
Split some of the logic out of the configure and build recipes into
makefile definitions, just to simplifiy things a little.  There should
be no functional as a result of this split.

I've also added a call to strip, so that the configure line will be
better formed, instead of this:

    configure --target=   TARGET-NAME

we instead get this:

    configure --target=TARGET-NAME

I've also replaced a pair of empty quotes with a blank string where
the CFLAGS are concerned, so, for targets that don't specify any
CFLAGS, we previously added: "" to the configure line, this is now
gone.